Move enum files to Enums/
