+++ app/Actions/CreateArgument.php
@@ -5,7 +5,7 @@ namespace App\Actions;
 use App\Models\Argument;
 use App\Models\Rfc;
 use App\Models\User;
-use App\Models\VoteType;
+use App\Models\Enums\VoteType;
 use Illuminate\Support\Facades\DB;
 
 final readonly class CreateArgument
