+++ app/Actions/RemoveReputation.php
@@ -2,7 +2,7 @@
 
 namespace App\Actions;
 
-use App\Models\ReputationType;
+use App\Models\Enums\ReputationType;
 use App\Models\User;
 
 final readonly class RemoveReputation
