+++ app/Http/Livewire/Enums/Events.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Livewire;
+namespace App\Http\Livewire\Enums;
 
 enum Events: string
 {
