+++ app/Models/Argument.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Models\Enums\VoteType;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
