+++ app/Models/Enums/ReputationType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Enums;
 
 enum ReputationType
 {
