+++ app/Models/Enums/VoteType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Enums;
 
 enum VoteType: string
 {
