+++ app/Models/User.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Models\Enums\UserFlair;
+use App\Models\Enums\VerificationRequestStatus;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
